// main/apps/tms/src/types/api/deviceDeployment.ts
export interface DeviceLatestDeploymentRequest {
  groupId: string
  siteId: string
  taskflowId: number
  robotId: string[]
}
export interface DeviceLatestDeploymentResponse {
  content: Content[]
  page: number
  size: number
  totalElements: number
  totalPages: number
  hasNext: true
  hasPrev: true
}

export interface Content {
  robotId: string
  deployments: Deployment[]
}

export interface Deployment {
  id: number
  groupId: string | null
  siteId: string | null
  deploymentId: number
  robotId: string
  taskFlowId: number
  taskFlowSnapshotId: number
  taskFlowVersion: number
  status: string
  retryCount: number
  errorMessage: string
  description: string
  deployedBy: string
  createdAt: string
  completedAt: string
  updatedAt: string
}